stage 8 - iteration (move to 64 compiler + some bugfixes). runtime is broken for now. Except for main menu

+++ new_game/src/engine/graphics/graphics_engine/backend_opengl/game/core.cpp
@@ -2114,7 +2114,7 @@ SLONG ClearDisplay(UBYTE r, UBYTE g, UBYTE b)
     return 0;
 }
 
-void SetLastClumpfile(char*, unsigned int) {}
+void SetLastClumpfile(char*, size_t) {}
 
 void* LockWorkScreen() { return s_work_screen_buf; }
 void UnlockWorkScreen() {}
